修改项目

+++ controllers/cert_cate.go
@@ -34,12 +34,14 @@ func GetCert_CateByID(c *gin.Context) {
 func CreateCert_Cate(c *gin.Context) {
 	var cert_cate *models.Cert_cate
 	c.BindJSON(&cert_cate)
-	err := models.AddCert_Cate(cert_cate)
-	if err != nil {
-		c.AbortWithStatus(404)
-		fmt.Println(err)
-	} else {
-		c.JSON(200, cert_cate)
+	if models.ExistsCert_CateById(int64(cert_cate.Id)) == false {
+		err := models.AddCert_Cate(cert_cate)
+		if err != nil {
+			c.AbortWithStatus(404)
+			fmt.Println(err)
+		} else {
+			c.JSON(200, cert_cate)
+		}
 	}
 }
 
@@ -68,10 +70,12 @@ func UpdateCert_Cate(c *gin.Context) {
 func DeleteCert_Cate(c *gin.Context) {
 	Id := c.Param("id")
 	id, _ := strconv.ParseInt(Id, 10, 64)
-	err := models.DeleteCert_Cate(id)
-	if err != nil {
-		c.AbortWithStatus(404)
-		fmt.Println(err)
+	if models.ExistsCert_CateById(id) == true {
+		err := models.DeleteCert_Cate(id)
+		if err != nil {
+			c.AbortWithStatus(404)
+			fmt.Println(err)
+		}
+		c.JSON(200, gin.H{"id #" + Id: "deleted"})
 	}
-	c.JSON(200, gin.H{"id #" + Id: "deleted"})
 }
