+++ controllers/cert_save.go
@@ -34,12 +34,14 @@ func GetCert_SaveByID(c *gin.Context) {
 func CreateCert_Save(c *gin.Context) {
 	var cert_save *models.Cert_save
 	c.BindJSON(&cert_save)
-	err := models.AddCert_Save(cert_save)
-	if err != nil {
-		c.AbortWithStatus(404)
-		fmt.Println(err)
-	} else {
-		c.JSON(200, cert_save)
+	if models.ExistsCert_SaveById(int64(cert_save.Id)) == false {
+		err := models.AddCert_Save(cert_save)
+		if err != nil {
+			c.AbortWithStatus(404)
+			fmt.Println(err)
+		} else {
+			c.JSON(200, cert_save)
+		}
 	}
 }
 
@@ -84,10 +86,12 @@ func UpdateCert_Save(c *gin.Context) {
 func DeleteCert_Save(c *gin.Context) {
 	Id := c.Param("id")
 	id, _ := strconv.ParseInt(Id, 10, 64)
-	err := models.DeleteCert_Save(id)
-	if err != nil {
-		c.AbortWithStatus(404)
-		fmt.Println(err)
+	if models.ExistsCert_SaveById(id) == true {
+		err := models.DeleteCert_Save(id)
+		if err != nil {
+			c.AbortWithStatus(404)
+			fmt.Println(err)
+		}
+		c.JSON(200, gin.H{"id #" + Id: "deleted"})
 	}
-	c.JSON(200, gin.H{"id #" + Id: "deleted"})
 }
